[Refactor] remove the UpperCase in package
[Refactor] remove the UpperCase in package
