--- BOOT/product/src/main/java/com/coopang/product/application/request/productStock/ProductStockDto.java
@@ -1,10 +1,0 @@
-package com.coopang.product.application.request.productStock;
-
-import java.util.UUID;
-import lombok.Data;
-
-@Data
-public class ProductStockDto {
-    private int amount;
-    private UUID orderId;
-}
