+++ BOOT/product/src/main/java/com/coopang/product/application/request/productstock/ProductStockDto.java
@@ -1,0 +1,10 @@
+package com.coopang.product.application.request.productstock;
+
+import java.util.UUID;
+import lombok.Data;
+
+@Data
+public class ProductStockDto {
+    private int amount;
+    private UUID orderId;
+}
