--- BOOT/product/src/main/java/com/coopang/product/domain/entity/productStockHistory/ProductStockHistoryChangeType.java
@@ -1,6 +1,0 @@
-package com.coopang.product.domain.entity.productStockHistory;
-
-public enum ProductStockHistoryChangeType {
-    INCREASE,  // 재고 증가
-    DECREASE  // 재고 감소
-}
