+++ BOOT/product/src/main/java/com/coopang/product/domain/entity/productstock/ProductStockConverter.java
@@ -1,0 +1,18 @@
+package com.coopang.product.domain.entity.productstock;
+
+import jakarta.persistence.AttributeConverter;
+import jakarta.persistence.Converter;
+
+@Converter(autoApply = true)
+public class ProductStockConverter implements AttributeConverter<ProductStock, Integer>{
+
+    @Override
+    public Integer convertToDatabaseColumn(ProductStock productStock) {
+        return productStock.getValue();
+    }
+
+    @Override
+    public ProductStock convertToEntityAttribute(Integer dbData) {
+        return new ProductStock(dbData);
+    }
+}
