+++ BOOT/product/src/main/java/com/coopang/product/domain/entity/productstockhistory/ProductStockHistoryChangeType.java
@@ -1,0 +1,6 @@
+package com.coopang.product.domain.entity.productstockhistory;
+
+public enum ProductStockHistoryChangeType {
+    INCREASE,  // 재고 증가
+    DECREASE  // 재고 감소
+}
