+++ BOOT/product/src/main/java/com/coopang/product/domain/service/product/ProductDomainService.java
@@ -4,12 +4,6 @@ package com.coopang.product.domain.service.product;
 import com.coopang.product.application.request.product.ProductDto;
 import com.coopang.product.domain.entity.category.CategoryEntity;
 import com.coopang.product.domain.entity.product.ProductEntity;
-import com.coopang.product.domain.entity.productStock.ProductStockEntity;
-import com.coopang.product.domain.entity.productStockHistory.ProductStockHistoryChangeType;
-import com.coopang.product.domain.entity.productStockHistory.ProductStockHistoryEntity;
-import com.coopang.product.domain.entity.productStock.ProductStock;
-import com.coopang.product.domain.repository.category.CategoryRepository;
-import com.coopang.product.domain.repository.product.ProductRepository;
 import com.coopang.product.infrastructure.repository.category.CategoryJpaRepository;
 import com.coopang.product.infrastructure.repository.product.ProductJpaRepository;
 import lombok.RequiredArgsConstructor;
