+++ BOOT/product/src/main/java/com/coopang/product/infrastructure/configuration/ProductDataInitializer.java
@@ -3,15 +3,13 @@ package com.coopang.product.infrastructure.configuration;
 import com.coopang.product.application.request.product.ProductDto;
 import com.coopang.product.domain.entity.category.CategoryEntity;
 import com.coopang.product.domain.entity.product.ProductEntity;
-import com.coopang.product.domain.entity.productStock.ProductStock;
-import com.coopang.product.domain.entity.productStock.ProductStockEntity;
-import com.coopang.product.domain.entity.productStockHistory.ProductStockHistoryChangeType;
-import com.coopang.product.domain.entity.productStockHistory.ProductStockHistoryEntity;
-import com.coopang.product.domain.service.product.ProductDomainService;
+import com.coopang.product.domain.entity.productstock.ProductStock;
+import com.coopang.product.domain.entity.productstock.ProductStockEntity;
+import com.coopang.product.domain.entity.productstockhistory.ProductStockHistoryChangeType;
+import com.coopang.product.domain.entity.productstockhistory.ProductStockHistoryEntity;
 import com.coopang.product.infrastructure.repository.category.CategoryJpaRepository;
 import com.coopang.product.infrastructure.repository.product.ProductJpaRepository;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
