+++ BOOT/product/src/main/java/com/coopang/product/infrastructure/repository/productstock/ProductStockRepositoryCustom.java
@@ -1,0 +1,8 @@
+package com.coopang.product.infrastructure.repository.productstock;
+
+import com.coopang.product.domain.entity.productstock.ProductStockEntity;
+import java.util.UUID;
+
+public interface ProductStockRepositoryCustom {
+    ProductStockEntity findAndLockProductStock(UUID productId);
+}
