--- BOOT/product/src/main/java/com/coopang/product/presentation/request/productStockHistory/ProductStockHistorySearchConditionDto.java
@@ -1,14 +1,0 @@
-package com.coopang.product.presentation.request.productStockHistory;
-
-import com.coopang.product.domain.entity.productStockHistory.ProductStockHistoryChangeType;
-import java.time.LocalDateTime;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class ProductStockHistorySearchConditionDto{
-    private ProductStockHistoryChangeType changeType;
-    private LocalDateTime startDate;
-    private LocalDateTime endDate;
-}
